requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
drf-spectacular>=0.27.2
django-filter>=24.2
django-cors-headers>=4.4.0
mysqlclient>=2.2.4
django-redis>=5.4.0
Pillow>=10.3.0
python-dotenv>=1.0.1
razorpay>=1.4.2
gunicorn>=22.0.0
whitenoise>=6.7.0
